Refactor StaxLoader hierarchy a bit.
